fix: allow using the ES imports syntax in the controllers
Convert the code to use the ES imports syntax to enable controllers to
use the ES imports syntax.

While this appears to be just a refactor, it actually allows the user
to use the ES imports syntax in the controllers so I'm considering it
a fix (or a feature, depending on how you look at it).

Fixes #12